fairweatherdays/requirements.txt
blinker==1.8.2
certifi==2019.11.28
chardet==3.0.4
click==8.1.7
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
flask==3.0.3
greenlet==1.1.2
idna==2.8
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.1.0
numpy==1.15.0
pep8==1.7.1
psutil==5.9.0
pycodestyle==2.8.0
PyGObject==3.36.0
python-apt==2.0.0+ubuntu0.20.4.7
python-dotenv==1.0.1
requests==2.22.0
requests-unixsocket==0.2.0
schedule==1.2.2
six==1.14.0
SQLAlchemy==1.4.32
ssh-import-id==5.10
unattended-upgrades==0.1
urllib3==1.25.8
werkzeug==3.0.6
zipp==3.20.2
